feat: add refund and settlement modules with financial events
- Add full refund system (controller, entity, DTOs, mapper, service)
- Add full settlement system (controller, entity, DTOs, mapper, service)
- Add FinancialEvent entity for auditing payment lifecycle events
- Add RefundReferenceGenerator and SettlementReferenceGenerator
- Add recordSettlement and recordRefund to LedgerService
- Add settlement/refund ledger entry creation in LedgerServiceImpl
- Add FinancialEvent and Settlement references to LedgerEntry
- Extend PaymentProcessorImpl with authorization engine and ledger recording
- Convert all amount fields from BigDecimal to Long across ledger, settlement, refund DTOs
- Fix Payment.refundedAmount NOT NULL migration issue for existing rows
- Add /api/v1/settlements/** and /api/v1/refunds/** to API key filter and security config

+++ src/main/java/com/moddynerd/transiq/payment/ledger/dto/MerchantBalanceResponse.java
@@ -1,10 +1,8 @@
 package com.moddynerd.transiq.payment.ledger.dto;
 
-import java.math.BigDecimal;
-
 public record MerchantBalanceResponse(
 
-        BigDecimal availableBalance,
+        Long availableBalance,
 
         String currency
